backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0
pytest>=7.0.0
requests>=2.28.0